+++ config/components.ts
@@ -66,6 +66,7 @@ export const components = {
   experimental: [
     '/components/what-is-experimental',
     '/components/auto-center',
+    '/components/date-picker-view',
     '/components/desense-text',
     '/components/dropdown',
     '/components/ellipsis',
