+++ src/components/date-picker-view/index.ts
@@ -1,0 +1,3 @@
+import { DatePickerView } from './date-picker-view'
+
+export default DatePickerView
