+++ src/components/date-picker/date-picker-utils.ts
@@ -1,0 +1,149 @@
+import dayjs from 'dayjs'
+import { generateIntArray } from '../../utils/generate-int-array'
+import { PickerColumn } from '../picker'
+
+export type Precision = 'year' | 'month' | 'day' | 'hour' | 'minute' | 'second'
+
+const precisionRankRecord: Record<Precision, number> = {
+  year: 0,
+  month: 1,
+  day: 2,
+  hour: 3,
+  minute: 4,
+  second: 5,
+}
+
+export function generateDatePickerColumns(
+  selected: string[],
+  min: Date,
+  max: Date,
+  precision: Precision
+) {
+  const ret: PickerColumn[] = []
+
+  const minYear = min.getFullYear()
+  const minMonth = min.getMonth() + 1
+  const minDay = min.getDate()
+  const minHour = min.getHours()
+  const minMinute = min.getMinutes()
+  const minSecond = min.getSeconds()
+
+  const maxYear = max.getFullYear()
+  const maxMonth = max.getMonth() + 1
+  const maxDay = max.getDate()
+  const maxHour = max.getHours()
+  const maxMinute = max.getMinutes()
+  const maxSecond = max.getSeconds()
+
+  const rank = precisionRankRecord[precision]
+
+  if (rank >= precisionRankRecord.year) {
+    const years: string[] = []
+    for (let i = minYear; i <= maxYear; i++) {
+      years.push(i.toString())
+    }
+    ret.push(years)
+  }
+
+  const firstDayInSelectedMonth = dayjs(
+    convertStringArrayToDate([selected[0], selected[1], '1'])
+  )
+  const selectedYear = parseInt(selected[0])
+  const selectedMonth = parseInt(selected[1])
+  const selectedDay = parseInt(selected[2])
+  const selectedHour = parseInt(selected[3])
+  const selectedMinute = parseInt(selected[4])
+
+  const isInMinYear = selectedYear === minYear
+  const isInMaxYear = selectedYear === maxYear
+  const isInMinMonth = isInMinYear && selectedMonth === minMonth
+  const isInMaxMonth = isInMaxYear && selectedMonth === maxMonth
+  const isInMinDay = isInMinMonth && selectedDay === minDay
+  const isInMaxDay = isInMaxMonth && selectedDay === maxDay
+  const isInMinHour = isInMinDay && selectedHour === minHour
+  const isInMaxHour = isInMaxDay && selectedHour === maxHour
+  const isInMinMinute = isInMinHour && selectedMinute === minMinute
+  const isInMaxMinute = isInMaxHour && selectedMinute === maxMinute
+
+  if (rank >= precisionRankRecord.month) {
+    const lower = isInMinYear ? minMonth : 1
+    const upper = isInMaxYear ? maxMonth : 12
+    const months = generateIntArray(lower, upper)
+    ret.push(months.map(v => v.toString()))
+  }
+  if (rank >= precisionRankRecord.day) {
+    const lower = isInMinMonth ? minDay : 1
+    const upper = isInMaxMonth ? maxDay : firstDayInSelectedMonth.daysInMonth()
+    const days = generateIntArray(lower, upper)
+    ret.push(days.map(v => v.toString()))
+  }
+  if (rank >= precisionRankRecord.hour) {
+    const lower = isInMinDay ? minHour : 0
+    const upper = isInMaxDay ? maxHour : 23
+    const hours = generateIntArray(lower, upper)
+    ret.push(
+      hours.map(v => ({
+        label: ('0' + v.toString()).slice(-2),
+        value: v.toString(),
+      }))
+    )
+  }
+  if (rank >= precisionRankRecord.minute) {
+    const lower = isInMinHour ? minMinute : 0
+    const upper = isInMaxHour ? maxMinute : 59
+    const minutes = generateIntArray(lower, upper)
+    ret.push(
+      minutes.map(v => ({
+        label: ('0' + v.toString()).slice(-2),
+        value: v.toString(),
+      }))
+    )
+  }
+  if (rank >= precisionRankRecord.second) {
+    const lower = isInMinMinute ? minSecond : 0
+    const upper = isInMaxMinute ? maxSecond : 59
+    const seconds = generateIntArray(lower, upper)
+    ret.push(
+      seconds.map(v => ({
+        label: ('0' + v.toString()).slice(-2),
+        value: v.toString(),
+      }))
+    )
+  }
+
+  return ret
+}
+
+export function convertDateToStringArray(
+  date: Date | undefined | null
+): string[] {
+  if (!date) return []
+  return [
+    date.getFullYear().toString(),
+    (date.getMonth() + 1).toString(),
+    date.getDate().toString(),
+    date.getHours().toString(),
+    date.getMinutes().toString(),
+    date.getSeconds().toString(),
+  ]
+}
+
+export function convertStringArrayToDate(
+  value: (string | null | undefined)[]
+): Date | null {
+  if (value.length === 0) return null
+  const yearString = value[0] ?? '1900'
+  const monthString = value[1] ?? '1'
+  const dateString = value[2] ?? '1'
+  const hourString = value[3] ?? '0'
+  const minuteString = value[4] ?? '0'
+  const secondString = value[5] ?? '0'
+  return new Date(
+    parseInt(yearString),
+    parseInt(monthString) - 1,
+    parseInt(dateString),
+    parseInt(hourString),
+    parseInt(minuteString),
+    parseInt(secondString)
+  )
+}
