+++ src/index.ts
@@ -13,6 +13,7 @@ export { default as Checkbox } from './components/checkbox'
 export { default as Collapse } from './components/collapse'
 export { default as ConfigProvider } from './components/config-provider'
 export { default as DatePicker } from './components/date-picker'
+export { default as DatePickerView } from './components/date-picker-view'
 export { default as DesenseText } from './components/desense-text'
 export { default as Dialog } from './components/dialog'
 export { default as Divider } from './components/divider'
